requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.10.0
python-dotenv==1.0.0
Pillow==10.2.0
whitenoise==6.6.0
gunicorn==21.2.0
